+++ cli/src/test/scala/co/topl/brambl/cli/ParamsWalletModuleTest.scala
@@ -5,7 +5,6 @@ import scopt.OParser
 
 import java.nio.file.Files
 import java.nio.file.Paths
-import java.nio.file.Path
 
 class ParamsWalletModuleTest extends FunSuite {